REF: misplaced repr tests (#32736)
