+++ pandas/tests/frame/test_timeseries.py
@@ -20,11 +20,6 @@ class TestDataFrameTimeSeriesMethods:
         df["A"] = rng
         assert np.issubdtype(df["A"].dtype, np.dtype("M8[ns]"))
 
-    def test_frame_datetime64_pre1900_repr(self):
-        df = DataFrame({"year": date_range("1/1/1700", periods=50, freq="A-DEC")})
-        # it works!
-        repr(df)
-
     def test_frame_append_datetime64_col_other_units(self):
         n = 100
