+++ pandas/tests/series/test_timeseries.py
@@ -2,8 +2,6 @@ from io import StringIO
 
 import numpy as np
 
-from pandas._libs.tslib import iNaT
-
 import pandas as pd
 from pandas import DataFrame, DatetimeIndex, Series, date_range, timedelta_range
 import pandas._testing as tm
@@ -88,19 +86,6 @@ class TestTimeSeries:
         series = Series(dates)
         assert np.issubdtype(series.dtype, np.dtype("M8[ns]"))
 
-    def test_series_repr_nat(self):
-        series = Series([0, 1000, 2000, iNaT], dtype="M8[ns]")
-
-        result = repr(series)
-        expected = (
-            "0   1970-01-01 00:00:00.000000\n"
-            "1   1970-01-01 00:00:00.000001\n"
-            "2   1970-01-01 00:00:00.000002\n"
-            "3                          NaT\n"
-            "dtype: datetime64[ns]"
-        )
-        assert result == expected
-
     def test_promote_datetime_date(self):
         rng = date_range("1/1/2000", periods=20)
         ts = Series(np.random.randn(20), index=rng)
@@ -123,12 +108,6 @@ class TestTimeSeries:
         result = rng.get_indexer(ts2.index)
         expected = rng.get_indexer(ts_slice.index)
         tm.assert_numpy_array_equal(result, expected)
-
-    def test_format_pre_1900_dates(self):
-        rng = date_range("1/1/1850", "1/1/1950", freq="A-DEC")
-        rng.format()
-        ts = Series(1, index=rng)
-        repr(ts)
 
     def test_groupby_count_dateparseerror(self):
         dr = date_range(start="1/1/2012", freq="5min", periods=10)
